app/requirements.txt
Flask==3.0.3
requests==2.27.1
google-auth==2.33.0
